Move auth handler to own file
